+++ tests/test_module_ilrmsdmatrix.py
@@ -113,63 +113,3 @@ def test_ilrmsdmatrix_init(ilrmsdmatrix):
     assert ilrmsdmatrix._origignal_config_file == ilrmsd_pars
     assert type(ilrmsdmatrix.params) == dict
     assert len(ilrmsdmatrix.params) != 0
-
-
-def test_ilrmsdmatrix_run(ilrmsdmatrix, mocker):
-    """Test ilrmsdmatrix run method."""
-    ilrmsdmatrix.previous_io = MockPreviousIO(path=ilrmsdmatrix.path)
-    mocker.patch(
-        "haddock.modules.BaseHaddockModule.export_io_models",
-        return_value=None,
-        )
-    ilrmsdmatrix.run()
-    assert Path(ilrmsdmatrix.path, "ilrmsd.matrix").exists()
-    assert Path(ilrmsdmatrix.path, "receptor_contacts.con").exists()
-    with open(Path(ilrmsdmatrix.path, "ilrmsd.matrix")) as f:
-        assert f.readline() == f"1 2 16.715{os.linesep}"
-    with open(Path(ilrmsdmatrix.path, "receptor_contacts.con")) as f:
-        lines = f.readlines()
-        assert lines[0] == f"A 37 38 39 40 43 44 45 69 71 72 75 90 93 94 96 132{os.linesep}"  # noqa : E501
-        assert lines[1] == f"B 10 11 12 16 17 48 51 52 53 54 56 57{os.linesep}"
-
-def test_ilrmsdmatrix_run_swappedchains(ilrmsdmatrix, mocker):
-    """Test ilrmsdmatrix run method."""
-    ilrmsdmatrix.previous_io = MockPreviousIO(path=ilrmsdmatrix.path)
-    ilrmsdmatrix.params["receptor_chain"] = "B"
-    ilrmsdmatrix.params["ligand_chains"] = ["A"]
-    mocker.patch(
-        "haddock.modules.BaseHaddockModule.export_io_models",
-        return_value=None,
-        )
-    ilrmsdmatrix.run()
-    assert Path(ilrmsdmatrix.path, "ilrmsd.matrix").exists()
-    assert Path(ilrmsdmatrix.path, "receptor_contacts.con").exists()
-    with open(Path(ilrmsdmatrix.path, "ilrmsd.matrix")) as f:
-        assert f.readline() == f"1 2 15.166{os.linesep}"
-    with open(Path(ilrmsdmatrix.path, "receptor_contacts.con")) as f:
-        lines = f.readlines()
-        assert lines[0] == f"B 10 11 12 16 17 48 51 52 53 54 56 57{os.linesep}"
-        assert lines[1] == f"A 37 38 39 40 43 44 45 69 71 72 75 90 93 94 96 132{os.linesep}"  # noqa : E501
-
-
-class MockPreviousIO:
-    """Mock previous IO class."""
-
-    def __init__(self, path):
-        self.path = path
-
-    def retrieve_models(self, individualize: bool = False):
-        """Retrieve models."""
-        shutil.copy(
-            Path(golden_data, "protprot_complex_1.pdb"),
-            Path(".", "protprot_complex_1.pdb"),
-            )
-        shutil.copy(
-            Path(golden_data, "protprot_complex_2.pdb"),
-            Path(".", "protprot_complex_2.pdb"),
-            )
-        model_list = [
-            PDBFile(file_name="protprot_complex_1.pdb", path="."),
-            PDBFile(file_name="protprot_complex_2.pdb", path="."),
-            ]
-        return model_list
